Complete 04-flex-information
I completed 04-flex-information succesfully. My design matches the desired spec, but a problem I'm facing rn is that my solutions are a bit bloated. I didn't have to turn the body into a flexbox (text-align would have sufficed) but I did; I didn't need the gap (margin-bottom would have worked instead). I need to determine if this is an actual concern or not.